Almost Done

+++ src/components/post-form/PostForm.jsx
@@ -19,8 +19,8 @@ export default function PostForm({ post }) {
     const userData = useSelector((state) => state.auth.userData);
 
     const submit = async (data) => {
-        if (post) {
-            const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
+        if (post && post.$id) {
+            const file = data.image?.[0] ? await appwriteService.uploadFile(data.image[0]) : null;
 
             if (file) {
                 appwriteService.deleteFile(post.featuredImage);
@@ -28,19 +28,21 @@ export default function PostForm({ post }) {
 
             const dbPost = await appwriteService.updatePost(post.$id, {
                 ...data,
-                featuredImage: file ? file.$id : undefined,
+                featuredImage: file ? file.$id : post.featuredImage,
             });
 
             if (dbPost) {
                 navigate(`/post/${dbPost.$id}`);
             }
         } else {
-            const file = await appwriteService.uploadFile(data.image[0]);
+            const file = data.image?.[0] && (await appwriteService.uploadFile(data.image[0]));
 
             if (file) {
-                const fileId = file.$id;
-                data.featuredImage = fileId;
-                const dbPost = await appwriteService.createPost({ ...data, userId: userData.$id });
+                data.featuredImage = file.$id;
+                const dbPost = await appwriteService.createPost({
+                    ...data,
+                    userId: userData.$id,
+                });
 
                 if (dbPost) {
                     navigate(`/post/${dbPost.$id}`);
@@ -48,6 +50,7 @@ export default function PostForm({ post }) {
             }
         }
     };
+
 
     const slugTransform = useCallback((value) => {
         if (value && typeof value === "string")
